Create strategy for transactions process
