+++ blockchain-watcher/src/domain/actions/evm/GetEvmTransactions.ts
@@ -1,8 +1,9 @@
-import { EvmBlock, EvmTransaction, ReceiptTransaction } from "../../entities";
+import { DefaultTransactions } from "./strategy/DefaultTransactions";
 import { EvmBlockRepository } from "../../repositories";
+import { NFTTransactions } from "./strategy/NFTTransactions";
+import { EvmTransaction } from "../../entities";
 import { GetEvmOpts } from "./PollEvm";
 import winston from "winston";
-import { GetTransactionsStrategy } from "./strategy/GetTransactionsStrategy";
 
 export class GetEvmTransactions {
   private readonly blockRepo: EvmBlockRepository;
@@ -29,13 +30,23 @@ export class GetEvmTransactions {
       `[${chain}][exec] Processing blocks [fromBlock: ${fromBlock} - toBlock: ${toBlock}]`
     );
 
-    const populatedTransactions = await new GetTransactionsStrategy(
-      this.blockRepo,
-      fromBlock,
-      toBlock,
-      chain,
-      opts
-    ).execute();
+    let populatedTransactions: EvmTransaction[] = [];
+
+    const processors = [
+      new DefaultTransactions(this.blockRepo, fromBlock, toBlock, chain, opts),
+      new NFTTransactions(this.blockRepo, fromBlock, toBlock, chain, opts),
+    ];
+
+    for (const filter of opts.filters!) {
+      for (const process of processors) {
+        const normalizeFilter = this.normalizeFilter(filter);
+
+        if (process.apply(normalizeFilter.topics)) {
+          const transaction = await process.execute(normalizeFilter);
+          populatedTransactions.push(...transaction);
+        }
+      }
+    }
 
     this.logger.info(
       `[${chain}][exec] Got ${populatedTransactions?.length} transactions to process [fromBlock: ${fromBlock} - toBlock: ${toBlock}]`
@@ -43,7 +54,24 @@ export class GetEvmTransactions {
 
     return populatedTransactions;
   }
+
+  private normalizeFilter(filter: Filter): Filter {
+    return {
+      addresses: filter.addresses.map((address) => address.toLowerCase()),
+      topics: filter.topics.map((topic) => topic.toLowerCase()),
+    };
+  }
 }
+
+export interface GetTransactions {
+  apply(topic: string[]): boolean;
+  execute(filter: Filter): Promise<EvmTransaction[]>;
+}
+
+export type Filter = {
+  addresses: string[];
+  topics: string[];
+};
 
 type Range = {
   fromBlock: bigint;
