+++ blockchain-watcher/src/domain/actions/evm/strategy/NFTTransactions.ts
@@ -1,5 +1,5 @@
 import { EvmBlock, EvmTransaction, ReceiptTransaction } from "../../../entities";
-import { Filter, GetTransactions } from "./GetTransactionsStrategy";
+import { GetTransactions, Filter } from "../GetEvmTransactions";
 import { EvmBlockRepository } from "../../../repositories";
 import { GetEvmOpts } from "../PollEvm";
